[Fix] 사용자 탈퇴 시 리뷰 사용자 필드명 수정

+++ Server/user/service/user.service.js
@@ -227,10 +227,10 @@ const cleanupUserData = async (userId) => {
       const deletedChatLogs = await ChatLog.deleteMany({ userId });
 
       // 리뷰 데이터 정리
-      const deletedReviews = await Review.deleteMany({ userId });
+      const deletedReviews = await Review.deleteMany({ author: userId });
 
       // Refresh Token 정리
-      const deletedTokens = await RefreshToken.deleteMany({ author: userId });
+      const deletedTokens = await RefreshToken.deleteMany({ user_id: userId });
 
       // S3에 업로드된 사용자 이미지 삭제
       await deleteUserImagesFromS3(userId);
